requirements.txt
python-gitlab>=1.13.0
treelib>=1.5.5